+++ CHANGES.md
@@ -5,6 +5,7 @@
 -   feat(cmd:release): use custom registry for each release profile.
 -   feat(cmd:release): allow `*` for any version of dependencies.
 -   fix(cmd:bootstrap): should install local devDependencies.
+-   fix(cmd:bootstrap): added `npm ci` supports.
 -   fix(cmd:recall): use registry from release config.
 
 ## v0.3.4